// src/app/shared/models/master-data-types.ts
export enum MasterDataTypes {
    SUBSTATION = 1,
    LINE = 14,
    HVDC = 15,
    BUSREACTOR = 4,
    LINEREACTOR = 5,
    GENERATINGSTATION = 7,
    GENERATINGUNIT = 8,
    TRANSFORMER = 9,
    FSC = 10,
    SVC = 11,
    TCSC = 12,
    HVDCPOLE = 13,
    BUS = 16,
    BAY = 25,
    AUTORECLOSE = 26,
    STATCOM = 11,
    FIRSTTIMECHARGE = 18
}